Work
Former-commit-id: 30175e35a7d0e20b65933b38e45e473d718c25dc

+++ pkg/txt2imgC/sample.go
@@ -15,7 +15,7 @@ func sample(
 	minCfg float32,
 	cfgScale float32,
 	method sampleMethodT,
-	sigmas []float32,
+	sigmas []float64,
 	startMergeStep int,
 	cID *ggml.Tensor,
 	cVecID *ggml.Tensor,
